fix: `queue add --wait` should use notification when available
Also:
- refactor cleanups to consolidate s3 waiting logic in the existing wait.go file
- simplify the existing polling logic to call Lsf on the specific file we are looking for
