+++ pkg/runtime/queue/s3.go
@@ -13,7 +13,6 @@ import (
 	"time"
 
 	"github.com/minio/minio-go/v7"
-	"github.com/minio/minio-go/v7/pkg/notification"
 	"golang.org/x/sync/errgroup"
 )
 
@@ -227,23 +226,6 @@ func (s3 S3Client) retryOnError(err error) bool {
 	return true
 }
 
-func (s3 S3Client) waitForBucket(bucket string) error {
-	// TODO use notifications
-	for {
-		exists, err := s3.BucketExists(bucket)
-		if err != nil {
-			return err
-		} else if !exists {
-			fmt.Fprintf(os.Stderr, "Waiting for bucket %s\n", bucket)
-			time.Sleep(1 * time.Second)
-		} else {
-			break
-		}
-	}
-
-	return nil
-}
-
 // This will wait for the s3 server to be reachable, but will not wait
 // for the bucket to exist
 func (s3 S3Client) BucketExists(bucket string) (bool, error) {
@@ -277,31 +259,4 @@ func (s3 S3Client) Mkdirp(bucket string) error {
 	}
 
 	return nil
-}
-
-func (s3 S3Client) WaitForEvent(bucket, object, event string) error {
-	defer s3.client.RemoveAllBucketNotification(s3.context, bucket)
-
-	suffix := ""
-	for notificationInfo := range s3.client.ListenBucketNotification(s3.context, bucket, object, suffix, []string{event}) {
-		if notificationInfo.Err != nil {
-			return notificationInfo.Err
-		}
-
-		break
-	}
-
-	return nil
-}
-
-func (s3 S3Client) WaitTillExists(bucket, object string) error {
-	return s3.WaitForEvent(bucket, object, "s3:ObjectCreated:*")
-}
-
-func (s3 S3Client) Listen(bucket, prefix, suffix string) <-chan notification.Info {
-	return s3.client.ListenBucketNotification(s3.context, bucket, prefix, suffix, []string{"s3:ObjectCreated:*"})
-}
-
-func (s3 S3Client) StopListening(bucket string) error {
-	return s3.client.RemoveAllBucketNotification(s3.context, bucket)
 }
